Feat: Add basic support for theme
